Moved pagerank tests to pagerank.rs and renamed struct to PageRankBuilder
